+++ megaByte/src/dao/ParticipantDBDAO.java
@@ -93,7 +93,7 @@ public class ParticipantDBDAO implements ParticipantDAO {
 	}// updateParticipant
 
 	@Override
-	public Participant getParticipant(int id) {
+	public Participant getParticipant(long id) {
 		Participant participant = null;
 		try {
 			Connection conn = _pool.getConnection();
